requirements.txt
# Larry Navigator v2.0 Dependencies
# 4D Problem Diagnosis System with Warm Educational UI

# Core Framework
streamlit>=1.28.0

# AI & LLM
google-genai>=1.0.0
google-generativeai>=0.4.1

# Web Research (optional but recommended)
tavily-python>=0.3.0

# Utilities
python-dotenv>=1.0.0

# Note: Neo4j and LangChain dependencies removed
# v2.0 uses only Gemini for simplified architecture
